+++ superset/sqllab/command.py
@@ -44,13 +44,14 @@ from superset.exceptions import (
 )
 from superset.jinja_context import BaseTemplateProcessor, get_template_processor
 from superset.models.core import Database
-from superset.models.sql_lab import LimitingFactor, Query
+from superset.models.sql_lab import Query
 from superset.queries.dao import QueryDAO
 from superset.sqllab.command_status import SqlJsonExecutionStatus
+from superset.sqllab.limiting_factor import LimitingFactor
+from superset.sqllab.utils import apply_display_max_row_configuration_if_require
 from superset.utils import core as utils
 from superset.utils.dates import now_as_float
 from superset.utils.sqllab_execution_context import SqlJsonExecutionContext
-from superset.sqllab.utils import apply_display_max_row_configuration_if_require
 
 config = app.config
 logger = logging.getLogger(__name__)
@@ -397,7 +398,9 @@ class ExecuteSqlCommand(BaseCommand):
     ) -> str:
         display_max_row = config["DISPLAY_MAX_ROW"]
         return json.dumps(
-            apply_display_max_row_configuration_if_require(execution_result, display_max_row),
+            apply_display_max_row_configuration_if_require(
+                execution_result, display_max_row
+            ),
             default=utils.pessimistic_json_iso_dttm_ser,
             ignore_nan=True,
             encoding=None,
